requirements.txt
fastapi
uvicorn
openai>=1.0.0
pyyaml
requests
python-dotenv
pydantic>=1.10
